compose: Replace autosize with custom implementation.
Replace the `autosize` JS library with a CSS approach to handle
resizing for compose box and message editing box.

The new implementation places the textarea inside a CSS Grid container
with a hidden replica element.
This allows the browser's native layout engine to handle resizing
automatically which improves the performance drastically.

This implementation is adapted from
https://css-tricks.com/the-cleanest-trick-for-autogrowing-textareas/

Fixes #36546.